+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `preserve_order` feature backing spec maps with `IndexMap` to keep insertion order
 
 ## [0.1.3] 2025-06-28
 ### Changed